🚚(frontend) move widget debounce function
Move this util function from lib-video to lib-components because this
is generic code that will be used in both lib-videos and lib-classroom
