+++ src/frontend/packages/lib_components/src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './useResizer';
 export * from './XAPI';
 export * from './XAPI/DocumentXapiStatement';
 export * from './formatSizeErrorScale';
+export * from './debounce';
