+++ src/frontend/packages/lib_video/src/components/common/VideoWidgetProvider/widgets/DescriptionWidget/index.tsx
@@ -1,12 +1,16 @@
-import { Video, report, TextAreaInput, FoldableItem } from 'lib-components';
+import {
+  Video,
+  report,
+  TextAreaInput,
+  FoldableItem,
+  debounce,
+} from 'lib-components';
 import React, { useEffect, useRef, useState } from 'react';
 import { toast } from 'react-hot-toast';
 import { defineMessages, useIntl } from 'react-intl';
 
 import { useUpdateVideo } from '@lib-video/api/useUpdateVideo';
 import { useCurrentVideo } from '@lib-video/hooks/useCurrentVideo';
-
-import { debounce } from '../../utils/widgets';
 
 const messages = defineMessages({
   info: {
